% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPR.R
\name{getPR}
\alias{getPR}
\title{Query Landsat wrs2 grid by country name or location}
\usage{
getPR(location)
}
\arguments{
\item{location}{character (Usually the name of a country) or an object of class Spatial*}
}
\value{
A list of path, rows, PathRows, and a spatialPolygon object
}
\description{
Performs a spatial query on the wrs2 grid using the worldHires dataset or an object of class Spatial*
}
\details{
Check the \code{\link{getCountryList}} function to get the right spelling
for the polygon name you want to use for the query.
}
\examples{
pr <- getPR('France')
pr$PATH
pr$ROW
pr$PR
map('worldHires', 'France', exact=TRUE)
plot(pr$map, add=TRUE)
}
\author{
Loïc Dutrieux
}
\seealso{
\code{\link{getCountryList}}
}
\keyword{Landsat}
\keyword{worldHires}

