% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpModis2Landsat.R
\name{warpModis2Landsat}
\alias{warpModis2Landsat}
\title{Reprojects MODIS data using target proj and extent}
\usage{
warpModis2Landsat(target, resolution = 30, ModisInput, ModisSds, filename)
}
\arguments{
\item{target}{Character. Target raster filename. Full path. Can Also be a list of characters.}

\item{resolution}{Numeric, resolution to resample to.}

\item{ModisInput}{Character. Input modis file.}

\item{ModisSds}{Numeric If \code{ModisInput} is a hdf file, the subdataset number of the data.}

\item{filename}{Character. Filename of the resulting output file. Full path.}
}
\value{
Character string of a gdalwarp expression. Before copy-pasting it in a terminal, you may need to print it using quote=FALSE.
}
\description{
Creates a character to be used in a system call, or directly in a terminal; and perform reprojection/resampling/subsetting of MODIS data, using proj and extent from a second raster.
}
\author{
Loic Dutrieux
}

