% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCountryList}
\alias{getCountryList}
\title{Identify polygon name from the worldHires dataset}
\usage{
getCountryList(country)
}
\arguments{
\item{country}{character. Usually the name of a country.}
}
\value{
A character list
}
\description{
Check for names of polygons matching the character provided (usually the
name of a country). Usefull prior to performing a getPR operation for
identifying the polygon name to use in the query.
}
\details{
The first letter of polygon names is often capitalized in the worldHires
dataset.
}
\examples{
# PLot Texel island (an island on the coast of the Netherlands)
list <- getCountryList('Netherlands')
list
map('worldHires', list[7], exact=TRUE, fill=TRUE)
}
\author{
Loïc Dutrieux
}
\seealso{
\code{\link{getPR}}
}
\keyword{worldHires}

