% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPR}
\alias{getPR}
\title{Query Landsat wrs2 grid by country name}
\usage{
getPR(country)
}
\arguments{
\item{country}{character. Usually the name of a country}
}
\value{
A list of path, rows, PathRows, and a spatialPolygon object
}
\description{
Performs a spatial query on the wrs2 grid using the worldHires dataset.
}
\details{
Check the \code{\link{getCountryList}} function to get the right spelling
for the polygon name you want to use for the query.
}
\examples{
pr <- getPR('France')
pr$PATH
pr$ROW
pr$PR
map('worldHires', 'France', exact=TRUE)
plot(pr$map, add=TRUE)
}
\author{
Loïc Dutrieux
}
\seealso{
\code{\link{getCountryList}}
}
\keyword{Landsat}
\keyword{worldHires}

