% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unpackVCF.R
\name{unpackVCF}
\alias{unpackVCF}
\title{Uncompress the .gz archives of the vcf data}
\usage{
unpackVCF(pr, year, searchDir, dir = NULL, mc.cores = 1)
}
\arguments{
\item{pr}{List or numeric list. Classically the returned object from
\code{\link{getPR}}.}

\item{year}{Numeric or list (i.e.: c(2000, 2005) or c(19902000, 20002005))}

\item{searchDir}{character. Root directory of where the archive are stored}

\item{dir}{character or \code{NULL}, if \code{NULL} (default), data are
unpacked in the directories containing the archive. If a directory is
provided, all data are unpacked in that directory.}

\item{mc.cores}{numeric. For parallel implementation, number of workers.}
}
\value{
A list of filenames and potential warning/error messages
}
\description{
Recursively searches through the directories and unpack the queried
archives.
}
\section{Warning }{
 For parallel implementation, see warning section of
\code{\link{mclapply}}
}
\examples{
\dontrun{
pr <- getPR('Belize')
pr
dir = tempdir()
downloadPR(pr, year=2000, dir=dir)
unpackVCF(pr=pr, year=2000, searchDir=dir, dir=sprintf('\%s/\%s',dir,'extract/'))
list.files(sprintf('\%s/\%s',dir,'extract/'))

}
}
\author{
Loic Dutrieux
}
\keyword{.gz}
\keyword{Landsat}

